Moved controller helpers to better place
